🔨 [entrypoints] Move function `extra_context_callback` to `cookiecutter`
