+++ src/cutty/entrypoints/cli/link.py
@@ -4,7 +4,7 @@ from typing import Optional
 
 import click
 
-from cutty.entrypoints.cli.create import extra_context_callback
+from cutty.entrypoints.cli.cookiecutter import extra_context_callback
 from cutty.services.link import link as service_link
 from cutty.templates.domain.bindings import Binding
